Offline Chat App iOS
Application is based on Offline Chat system where iOS users can
interact with
other iOS users who are nearby without using internet.
Whenever any user is using the app this app would connect to other
users which would would allow users to chat with each other. This app
allows multiple users to connect each other if they are nearby to range
of Wifi & Bluetooth.
Currently the app is limited to basic chat only to a group.
App Features:
MultipeerConnectivity: This framework supports the discovery of
services provided by nearby devices and supports communicating with
those services through message-based data, streaming data, and
resources (such as files). In iOS, the framework uses infrastructure
Wi-Fi networks, peer-to-peer Wi-Fi, and Bluetooth personal area
networks for the underlying transport. In macOS and tvOS, it uses
infrastructure Wi-Fi, peer-to-peer Wi-Fi, and Ethernet.

Mobile 1
This is device 1 which shows the connection which it has established
with other users showing the names of those devices.

Mobile 2
This is device 2 which shows the connection which it has established
with other users showing the names of those devices.

Mobile 3
This is device 3 which shows the connection which it has established
with other users showing the names of those devices.

In the first 3 screenshot from 3 different devices that they have
established the connection between them which can be seen from
connection label.

Mobile 1
This is device 1 which shows the messages which it has exchanged with
other users with the names of those devices who has sent the message.

Mobile 2
This is device 2 which shows the messages which it has exchanged with
other users with the names of those devices who has sent the message.

Mobile 3

This is device 3 which shows the messages which it has exchanged with
other users with the names of those devices who has sent the message.

In the second screenshot from the all 3 devices when everyone sent the
message everyone can see and respond the messages.